fix: issues with shipstation
